Fix ruff lint errors and configure ruff to ignore stylistic-only rules
Auto-fixed 233 lint issues (import sorting, typing.List -> list, etc.)
that were blocking CI. Also fixes a real bug in
tests/test_visual.py::test_near_black_high_variance_not_flagged, which
referenced an undefined `is_black` (the test body was incomplete and
would raise NameError). Added [tool.ruff.lint] ignores for rules that
flag intentional patterns (defensive except Exception, FastAPI's
File(...) default-arg idiom, sync file I/O in upload handler, unused
tuple-unpack in tests).

+++ backend/pipeline/frames.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
+
 import subprocess
 from pathlib import Path
+
 import imageio_ffmpeg
 
 from backend.pipeline.workspace import Workspace
@@ -22,7 +24,7 @@ def sample_frames(
     scale_expr = (
         f"scale='if(gt(iw,ih),{longest_side},-2)':'if(gt(iw,ih),-2,{longest_side})'"
     )
-    qv = max(2, min(31, int(round((100 - jpeg_quality) / 3))))
+    qv = max(2, min(31, round((100 - jpeg_quality) / 3)))
 
     cmd = [
         _FFMPEG_EXE,
